% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{contingency_matrix}
\alias{contingency_matrix}
\title{Contingency Table}
\usage{
contingency_matrix(M1, M2)
}
\arguments{
\item{M1, M2}{Two integer matrices of the same size.}
}
\description{
Contingency Table
}
