% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similR.R
\name{jsd}
\alias{jsd}
\title{Jensen-Shannon divergence of discrete probability distibutions JSD(p||q)}
\usage{
jsd(p, q)
}
\arguments{
\item{p}{first probability distribution}

\item{q}{second probability distribution}
}
\value{
JSD(p||q), a numeric value
}
\description{
Jensen-Shannon divergence of discrete probability distibutions JSD(p||q)
}
\details{
Assumes that both p and q are normalized (sum(p)=sum(q)=1) and that they contain no NA values. Calculated with base-e log, so the bounds are [0,log(2)].
}
