% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similR.R
\name{kld}
\alias{kld}
\title{Kullback-Leibler divergence between discrete probability distributions KLD(p||q)}
\usage{
kld(p, q)
}
\arguments{
\item{p}{first probability distribution ("to which"), a numeric vector, no NAs}

\item{q}{second probability distribution ("from which"), a numeric vector of same length as \code{p}, no NAs}
}
\value{
KLD(p||q), a numeric value
}
\description{
Kullback-Leibler divergence between discrete probability distributions KLD(p||q)
}
\details{
Assumes that both p and q are normalized (sum(p)=sum(q)=1). Note that by definition the Kullback-Leibler divergence is asymmetric: KLD(p||q) is not equal to KLD(q||p).
}
