% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/las.R
\name{las}
\alias{las}
\alias{LAS}
\alias{CS}
\alias{Concensus-Structure}
\alias{las.list}
\alias{las.matrix}
\title{Locally Aggregated Structures (LAS)}
\usage{
las(M, ..., rule, threshold = 0.5)

LAS(M, ..., rule, threshold = 0.5)

\method{las}{list}(M, ..., rule, threshold = 0.5)

\method{las}{matrix}(M, ..., rule, threshold = 0.5)
}
\arguments{
\item{M}{Either a list of matrices of size \code{n} (need not to be square), or
a single matrix of size \code{n} (see details).}

\item{...}{More matrices to be passed to the function.}

\item{rule}{Character. Rule used to create the LAS matrix. Can be
any of \code{"intersection"}, \code{"union"}, or \code{"threshold"} (see the \emph{Rule} section).}

\item{threshold}{Number of minimum agreements to set the tie to 1. This is used
for the case when \code{rule = "threshold"}.}
}
\description{
Generates LAS matrices using different criteria as described in Krackhardt (1990).
}
\section{Rule}{

When \code{rule = "interesection"} or \code{rule = "i"}, the union method is used.
A tie is marked as present iff both i and j agree on the existance if (i,j).


When \code{rule = "union"} or \code{rule = "u"}, the union method is used. In such case
a tie is marked as present if any of the individuals declares it.


When \code{rule = "threshold"} or \code{rule = "t"}, the threshold method is used in which
a tie is marked as present iff the proportion of individuals who agree on the
existance of that is greater than \code{threshold}. In its paper, Krackhardt calls
this \emph{Concensus Structure} (CS).
}

\examples{
data(powerset03)
x <- powerset03[1:3]

# It only exists if i and j declare it
las(x, rule="i")

# Same as above
las(x[[1]], x[[2]], x[[3]], rule="i")

# It exists if either i or j declare it
las(x, rule="u")

# Not enough agreement
las(x, rule="t", threshold=.9)
}
\references{
Krackhardt, D. (1987). Cognitive social structures. Social networks, 9(2), 109-134.
}
