% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerset.R
\docType{data}
\name{powerset03}
\alias{powerset03}
\alias{powerset04}
\title{Power sets of matrices of size 3 and 4}
\format{An object of class \code{list} of length 64.}
\usage{
powerset03

powerset04
}
\description{
Power sets of matrices of size 3 and 4
}
\keyword{datasets}
