% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/similarity.R
\name{similarity}
\alias{similarity}
\alias{Jaccard}
\alias{Tarwid}
\alias{Person-&-Heron}
\alias{Hamming}
\alias{Dennis}
\alias{Yuleq}
\alias{Michael}
\alias{Peirce}
\alias{distance}
\alias{similarity.list}
\alias{similarity.matrix}
\title{Similarity and Distance between pairs of binary matrices}
\usage{
similarity(M, ..., statistic, normalized = TRUE)

distance(M, ..., statistic, normalized = TRUE)

\method{similarity}{list}(M, ..., statistic, normalized = TRUE)

\method{similarity}{matrix}(M, ..., statistic, normalized = TRUE)
}
\arguments{
\item{M}{Either a list of matrices of size `n` (need not to be square), or
a single matrix of size `n` (see details).}

\item{...}{More matrices to be passed to the function.}

\item{statistic}{Character. Name of the similarity index to be using.}

\item{normalized}{Logical. When `TRUE` it returns the normalized hamming distance,
which ranges between 0 and 1 (currently only used in `statistic="hamming"`).}
}
\value{
A matrix of size `n*(n - 1)/2` by `length(statistic)`, where columns
1 and 2 indicate the id if `i` and `j`, and the reminder columns are the
corresponding distances/similarities.
}
\description{
Multiple measurements of similarity and distance between pairs of binary
matrices as listed in Choi et al (2010).
}
\details{
All of the available statistics are based on a 2x2 contingency matrix counting
matches and missmatches between each pair of matrices.

\tabular{llcc}{
     \tab   \tab R2 \tab   \cr
     \tab   \tab 1  \tab 0 \cr
  R1 \tab 1 \tab a  \tab b \cr
     \tab 0 \tab c  \tab d  
}

A complete list of the statistics available is available in the **Similarity**
and **Distance** sections.

`distance` is just an alias for `similarity`.
}
\section{Similarity}{

\itemize{
\item Jaccard (1): \code{"sjaccard"} or \code{"jaccard"}
}


\itemize{
\item Tarwid (54): \code{"starwid"} or \code{"tarwid"}.
}


\itemize{
\item Pearson & Heron 1 (54): \code{"sph1"} or \code{"ph1"} or \code{"s14"}. This is also known as S14 in
Gower and Legendre (1986).

In the case of the \code{S14} function, following Krackhardt's 1989:

\deqn{%
   \sqrt{\left(\frac{a}{(a + c)} - \frac{b}{(b + d)}\right)\times\left(\frac{a}{(a + b)} - \frac{c}{(c + d)}\right)}
   }{%
   S14 = [(a/(a + c) - b/(b + d))*(a/(a + b) - c/(c + d))]^(1/2)
   }

Which is an statistic lying between 0 and 1.
}


\itemize{
\item Dennis (44): \code{"sdennis"} or \code{"dennis"}
}


\itemize{
\item Yuleq (61): \code{"syuleq"}
}


\itemize{
\item Yuleq (63): \code{"syuleqw"}
}


\itemize{
\item Michael (68):  \code{"smichael"} or \code{"michael"}
}


\itemize{
\item Peirce (73): \code{"speirce"} or \code{"peirce"}
}
}

\section{Distance}{

\itemize{
\item Hamming (15): \code{"dhamming"} or \code{"hamming"}
}


\itemize{
\item Yuleq (62): \code{"dyuleq"}
}
}

\examples{
# Getting all networks of size 3
data(powerset03)
 
# We can compute it over the entire set
head(similarity(powerset03, statistic="s14"))

# Or over two pairs
head(similarity(powerset03[[1]], powerset03[[2]], powerset03[[3]], statistic="s14"))

# We can compute multiple distances at the same time
ans <- similarity(powerset03, statistic=c("hamming", "dennis", "jaccard"))
head(ans)

}
\references{
Choi, S. S., Cha, S. H., & Tappert, C. C. (2010). A survey of binary similarity
and distance measures. Journal of Systemics, Cybernetics and Informatics, 8(1), 43-48.

Krackhardt, D. (1990). Assessing the political landscape: Structure, cognition,
and power in organizations. Administrative science quarterly, 342-369.

Gower, J. C., & Legendre, P. (1986). Metric and Euclidean properties of
dissimilarity coefficients. Journal of classification, 3(1), 5-48.
}
