% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerset.R
\docType{data}
\name{statistics}
\alias{statistics}
\title{List of available statistics in \code{similarity}
The data object \code{statistics} contains a list of the available for
convenience.}
\format{A list of two character vectors.}
\usage{
statistics
}
\description{
List of available statistics in \code{similarity}
The data object \code{statistics} contains a list of the available for
convenience.
}
\examples{
data(statistics)
}
\seealso{
\link{similarity}
}
\keyword{datasets}
