\name{UniLitR-package}
\alias{UniLitR}
\alias{UniLitR-package}
\title{UniLitR: local recommendation of institutional scientific literature}
\description{
Infrastructure for verified local DuckDB data releases and a local Shiny
application used by the UniLitR recommendation system.
}
\docType{package}
\keyword{internal}
