% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerset.R
\docType{data}
\name{powerset03}
\alias{powerset03}
\alias{powerset04}
\title{Power sets of matrices of size 3 and 4}
\format{A List of integer matrices of size 3x3 (\code{powerset03}), and 4x4
(\code{powerset04}).}
\usage{
powerset03

powerset04
}
\description{
Power sets of matrices of size 3 and 4
}
\examples{
data(powerset03)
data(powerset04)
}
\keyword{datasets}
