% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/similarity.R
\name{similarity}
\alias{similarity}
\alias{Jaccard}
\alias{Faith}
\alias{Gower-&-Legendre}
\alias{Sized-Difference}
\alias{Shape-Difference}
\alias{Tarwid}
\alias{Person-&-Heron}
\alias{Hamming}
\alias{Mean-Manhattan}
\alias{Dennis}
\alias{Yuleq}
\alias{Michael}
\alias{Dispersion}
\alias{Hamann}
\alias{Goodman-&-Kruskal}
\alias{Anderberg}
\alias{Peirce}
\alias{similarity.list}
\alias{similarity.matrix}
\title{Similarity and Distance between pairs of binary matrices}
\usage{
similarity(M, ..., statistic, normalized = TRUE, firstonly = FALSE,
  exclude_j = FALSE)

\method{similarity}{list}(M, ..., statistic, normalized = TRUE,
  firstonly = FALSE, exclude_j = FALSE)

\method{similarity}{matrix}(M, ..., statistic, normalized = TRUE,
  firstonly = FALSE, exclude_j = FALSE)
}
\arguments{
\item{M}{Either a list of matrices of size \code{n} (need not to be square), or
a single matrix of size \code{n} (see details).}

\item{...}{More matrices to be passed to the function.}

\item{statistic}{Character. Name of the similarity index to be using.}

\item{normalized}{Logical. When \code{TRUE} it returns the normalized hamming distance,
which ranges between 0 and 1 (currently only used in \code{statistic="hamming"}).}

\item{firstonly}{Logical. When \code{TRUE}, the comparison is done as the first
matrix to all only.}

\item{exclude_j}{Logical. When \code{TRUE}, the comparison between matrices \code{i} and
\code{j} is done after the jth column and rows are removed from each.}
}
\value{
A matrix of size \code{n*(n - 1)/2} by \code{length(statistic)}, where columns
1 and 2 indicate the id if \code{i} and \code{j}, and the reminder columns are the
corresponding distances/similarities.
}
\description{
Multiple measurements of similarity and distance between pairs of binary
matrices as listed in Choi et al (2010).
}
\details{
All of the available statistics are based on a 2x2 contingency matrix counting
matches and missmatches between each pair of matrices (R1, R2).

\tabular{llcc}{
\tab   \tab R2 \tab   \cr
\tab   \tab 1  \tab 0 \cr
R1 \tab 1 \tab a  \tab b \cr
\tab 0 \tab c  \tab d
}

A complete list of the statistics available is available in the \strong{Similarity}
and \strong{Distance} sections.

\code{distance} is just an alias for \code{similarity}.
}
\section{Similarity}{

\itemize{
\item Jaccard (1): \code{"sjaccard"} or \code{"jaccard"}
}


\itemize{
\item Faith (10): \code{"sfaith"} or \code{"faith"}
}


\itemize{
\item Gower & Legendre (11): \code{"sgl"} or \code{"gl"}
}


\itemize{
\item Tarwid (54): \code{"starwid"} or \code{"tarwid"}.
}


\itemize{
\item Pearson & Heron 1 (54): \code{"sph1"} or \code{"ph1"} or \code{"s14"}. This is also known as S14 in
Gower and Legendre (1986).

In the case of the \code{S14} function, following Krackhardt's 1989:

\deqn{%
   \sqrt{\left(\frac{a}{(a + c)} - \frac{b}{(b + d)}\right)\times\left(\frac{a}{(a + b)} - \frac{c}{(c + d)}\right)}
   }{%
   S14 = [(a/(a + c) - b/(b + d))*(a/(a + b) - c/(c + d))]^(1/2)
   }

Which is an statistic lying between 0 and 1.
}


\itemize{
\item Mean Manhattan (20): \code{"dmh"} or \code{"mh"}
\deqn{%
     D_{Mean-manhattan} = \frac{b + c}{a + b + c + d}
   }{%
    dmh = (b + c)/(a + b + c + d)
   }
}


\itemize{
\item Dennis (44): \code{"sdennis"} or \code{"dennis"}
}


\itemize{
\item Yuleq (61): \code{"syuleq"}
}


\itemize{
\item Yuleq (63): \code{"syuleqw"}
}


\itemize{
\item Michael (68):  \code{"smichael"} or \code{"michael"}

\deqn{%
     S_{michael} = \frac{4(ad-bc)}{(a+d)^2 + (b+c)^2}
   }{%
     smichael = 4*(a*d - b*c)/[(a + d)^2 + (b + c)^2]
   }
}


\itemize{
\item Dispersion (66):  \code{"sdisp"} or \code{"disp"}

\deqn{%
     S_{Dispersion} = \frac{ad - bc}{(a + b + c + d)^2}
   }{%
     sdisp = [a*d - b*c]/(a + b + c + d)^2
   }
}


\itemize{
\item Hamann (67):  \code{"shamann"} or \code{"hamann"}

\deqn{%
     S_{Hamann} = \frac{(a + d) - (b + c)}{a + b + c + d}
   }{%
     shamann = [(a + d) - (b + c)](a + b + c + d)
   }
}


\itemize{
\item Goodman & Kruskal (69): \code{"sgk"} or \code{"gk"}

\deqn{%
     S_{Goodman&Kruskal} = \frac{\sigma - \sigma'}{2n - \sigma'} 
   }{%
     sgk = (s + s')/(2*n - s') 
   }

where
\eqn{\sigma = \max(a,b) + \max(c,d) + \max(a,c) + \max(b,d)}{
   s = max(a,b) + max(c,d) + max(a,c) + max(b,d)
   }, and
\eqn{\sigma' = \max(a + c, b + d) + \max(a + b, c + d)}{
   s' = max(a + c, b + d) + max(a + b, c + d)
   }
}


\itemize{
\item Anderberg (70): \code{"sanderberg"} or \code{"anderberg"}

\deqn{%
     S_{Anderberg} = \frac{\sigma - \sigma'}{2n} 
   }{%
     sanderberg = (s + s')/(2*n) 
   }

where \eqn{\sigma}{s} and \eqn{\sigma}{s'} are defined as in (69).
}


\itemize{
\item Peirce (73): \code{"speirce"} or \code{"peirce"}

\deqn{%
     S_{Peirce} = \frac{ab + bc}{ab + 2bc + cd}
   }{%
     speirce = (a*b + b*c)/(a*b + 2*b*c + c*d)
   }
}
}

\section{Distance}{

\itemize{
\item Sized Difference (24): \code{"dsd"} or \code{"sd"}
}


\itemize{
\item Shaped Difference (25): \code{"dsphd"} or \code{"sphd"}
}


\itemize{
\item Hamming (15): \code{"dhamming"} or \code{"hamming"}
}


\itemize{
\item Yuleq (62): \code{"dyuleq"}
}
}

\examples{
# Getting all networks of size 3
data(powerset03)
 
# We can compute it over the entire set
head(similarity(powerset03, statistic="s14"))

# Or over two pairs
head(similarity(powerset03[[1]], powerset03[[2]], powerset03[[3]], statistic="s14"))

# We can compute multiple distances at the same time
ans <- similarity(powerset03, statistic=c("hamming", "dennis", "jaccard"))
head(ans)

}
\references{
Choi, S. S., Cha, S. H., & Tappert, C. C. (2010). A survey of binary similarity
and distance measures. Journal of Systemics, Cybernetics and Informatics, 8(1), 43-48.

Krackhardt, D. (1990). Assessing the political landscape: Structure, cognition,
and power in organizations. Administrative science quarterly, 342-369.

Gower, J. C., & Legendre, P. (1986). Metric and Euclidean properties of
dissimilarity coefficients. Journal of classification, 3(1), 5-48.
}
\seealso{
The \link{statistics} object contains a list with the available statistics
for convenience.
}
