\name{similR-package}
\alias{similR}
\alias{similR-package}
\title{similR: local recommendation of institutional scientific literature}
\description{
Tools for preparing, versioning, validating, downloading, and using an institutional scientific literature database with a local Shiny application.
}
\docType{package}
\keyword{internal}
