% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similR.R
\name{similR}
\alias{similR}
\title{similR}
\usage{
similR(toks, vec = NULL, window_weights = 1/(1:5),
  word_vectors_size = 30, x_max = 10, n_iter = 30, ik = 100,
  clustering_algorithm = c("MacQueen", "Hartigan-Wong", "Lloyd",
  "Forgy"), clustering_itermax = 1000, similarity_method = "cosine",
  keep_vec = FALSE)
}
\arguments{
\item{toks}{tokens representing a document corpus}

\item{vec}{if NULL, will calculate the vector embedding of words. If this lengthy calculation needs to be skipped - pass the prepared matrix of vectors here (for instance, taken from the previous run of this function with keep_vec=TRUE).}

\item{window_weights}{weights of the window to use for co-occurrence of tokens.}

\item{word_vectors_size}{dimensionality of vector space}

\item{x_max}{max number of co-occurrences to use in the weighting function}

\item{n_iter}{number of GloVe iterations}

\item{ik}{initial number of clusters (the final one will be greater or equal than this)}

\item{clustering_algorithm}{as implemented in kmeans(), but note that 'Hartigan-Wong' fails for large data, so here we made 'MacQueen' the default.}

\item{clustering_itermax}{as implemented in kmeans()}

\item{similarity_method}{as implemented in quanteda::textstat_simil()}

\item{keep_vec}{whether to return the matrix of the word-vectors}
}
\value{
depending on keep_vec, either just the simil matrix/object, or a list with that and the matrix of the word-vectors
}
\description{
Document similarity based on word vectorization
}
